Rename h264 encoder
