+++ pkg/encoder/h264/encoder.go
@@ -1,4 +1,4 @@
-package h264encoder
+package h264
 
 import (
 	"bytes"
