+++ pkg/encoder/h264/x264.go
@@ -1,4 +1,4 @@
-package h264encoder
+package h264
 
 /*
 #include <stdlib.h>
